Adding roles and permissions

+++ routes/api.php
@@ -20,36 +20,38 @@ Route::middleware('guest')->group(function () {
 Route::middleware('auth:sanctum')->group(function () {
     Route::post('logout', [UserController::class, 'logout']);
 
-    Route::middleware('role:admin')->group(function () {});
-    Route::get('users', [UserController::class, 'index']);
-    Route::get('user/{id}', [UserController::class, 'show']);
-    Route::delete('user/{id}', [UserController::class, 'destroy']);
+    Route::get('users', [UserController::class, 'index'])->middleware('permission:all-users');
+    Route::get('user/{id}', [UserController::class, 'show'])->middleware('permission:find-user');
+    Route::delete('user/{id}', [UserController::class, 'destroy'])->middleware('permission:delete-user');
     Route::patch('user/{id}', [UserController::class, 'update']);
     Route::post('user/filter', [UserController::class, 'filter']);
     Route::post('user/search', [UserController::class, 'search']);
 
     //Route::resource('topic',TopicController::class);
     Route::prefix('topic')->group(function () {
-        Route::get('', [TopicController::class, 'index']);
-        Route::post('', [TopicController::class, 'store']);
+        Route::get('', [TopicController::class, 'index'])->middleware('permission:all-topics');
+        Route::post('', [TopicController::class, 'store'])->middleware('permission:create-topic');
         Route::get('/{id}', [TopicController::class, 'show']);
-        Route::post('/{id}', [TopicController::class, 'update']);
-        Route::delete('/{id}', [TopicController::class, 'destroy']);
+        Route::patch('/{id}', [TopicController::class, 'update'])->middleware('permission:update-topic');
+        Route::delete('/{id}', [TopicController::class, 'destroy'])->middleware('permission:delete-topic');
+        Route::post('/filter', [TopicController::class, 'filter']);
+        Route::post('/search', [TopicController::class, 'search']);
     });
 
-    Route::prefix('category')->group(function () {
+    Route::prefix('category')->middleware('role:Admin')->group(function () {
         Route::get('', [CategoryController::class, 'index']);
         Route::post('', [CategoryController::class, 'store']);
         Route::get('/{id}', [CategoryController::class, 'show']);
-        Route::post('/{id}', [CategoryController::class, 'update']);
+        Route::patch('/{id}', [CategoryController::class, 'update']);
         Route::delete('/{id}', [CategoryController::class, 'destroy']);
     });
 
     Route::prefix('response')->group(function () {
         Route::get('', [ResponseController::class, 'index']);
-        Route::post('', [ResponseController::class, 'store']);
+        Route::post('', [ResponseController::class, 'store'])->middleware('permission:create-response');
         Route::get('/{id}', [ResponseController::class, 'show']);
-        Route::post('/{id}', [ResponseController::class, 'update']);
-        Route::delete('/{id}', [ResponseController::class, 'destroy']);
+        Route::get('/topic/{id}', [ResponseController::class, 'topicResponses']);
+        Route::patch('/{id}', [ResponseController::class, 'update'])->middleware('permission:update-response');
+        Route::delete('/{id}', [ResponseController::class, 'destroy'])->middleware('permission:delete-response');
     });
 });
